// src/enums/StepRenseignerConventionRegionaleSpecifique.ts
enum StepRenseignerConventionRegionaleSpecifique {
    STEP_1 = 'Informations générales',
    STEP_2 = 'Objet et objectifs de la convention',
    STEP_3 = 'Consistance',
    STEP_4 = 'Domiciliation des contributions',
    STEP_5 = 'Contribution des partenaires',
    STEP_6 = 'Engagement du Ministère',
    STEP_7 = 'Mode de gouvernance',
    STEP_8 = 'Remarques et observations',
}

export default StepRenseignerConventionRegionaleSpecifique;